Access the shadowed firmware files and load them
The directory /lib/firmware/ti-connectivity is mounted from the HostOS and files copied by the apt-get install bb-wl18xx-firmware are still there but shadowed by the mount.

A possible workaround comes from http://superuser.com/questions/389255/is-it-possible-to-access-files-shadowed-by-a-mount. Mount at runtime (here the startup shell script) the /lib/firmware/ti-connectivity and copy back the files into the original directory.